# Week11/openapi_server/models/error.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Error(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, message=None, code=None):  # noqa: E501
        """Error - a model defined in OpenAPI

        :param message: The message of this Error.  # noqa: E501
        :type message: str
        :param code: The code of this Error.  # noqa: E501
        :type code: int
        """
        self.openapi_types = {
            'message': str,
            'code': int
        }

        self.attribute_map = {
            'message': 'message',
            'code': 'code'
        }

        self._message = message
        self._code = code

    @classmethod
    def from_dict(cls, dikt) -> 'Error':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Error of this Error.  # noqa: E501
        :rtype: Error
        """
        return util.deserialize_model(dikt, cls)

    @property
    def message(self) -> str:
        """Gets the message of this Error.

        Mô tả chi tiết về lỗi  # noqa: E501

        :return: The message of this Error.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this Error.

        Mô tả chi tiết về lỗi  # noqa: E501

        :param message: The message of this Error.
        :type message: str
        """
        if message is None:
            raise ValueError("Invalid value for `message`, must not be `None`")  # noqa: E501

        self._message = message

    @property
    def code(self) -> int:
        """Gets the code of this Error.

        Mã lỗi (tùy chọn)  # noqa: E501

        :return: The code of this Error.
        :rtype: int
        """
        return self._code

    @code.setter
    def code(self, code: int):
        """Sets the code of this Error.

        Mã lỗi (tùy chọn)  # noqa: E501

        :param code: The code of this Error.
        :type code: int
        """

        self._code = code