# Week11/openapi_server/models/product_input.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class ProductInput(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, price=None, description=None):  # noqa: E501
        """ProductInput - a model defined in OpenAPI

        :param name: The name of this ProductInput.  # noqa: E501
        :type name: str
        :param price: The price of this ProductInput.  # noqa: E501
        :type price: float
        :param description: The description of this ProductInput.  # noqa: E501
        :type description: str
        """
        self.openapi_types = {
            'name': str,
            'price': float,
            'description': str
        }

        self.attribute_map = {
            'name': 'name',
            'price': 'price',
            'description': 'description'
        }

        self._name = name
        self._price = price
        self._description = description

    @classmethod
    def from_dict(cls, dikt) -> 'ProductInput':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ProductInput of this ProductInput.  # noqa: E501
        :rtype: ProductInput
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this ProductInput.

        Tên của sản phẩm  # noqa: E501

        :return: The name of this ProductInput.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this ProductInput.

        Tên của sản phẩm  # noqa: E501

        :param name: The name of this ProductInput.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def price(self) -> float:
        """Gets the price of this ProductInput.

        Giá của sản phẩm  # noqa: E501

        :return: The price of this ProductInput.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this ProductInput.

        Giá của sản phẩm  # noqa: E501

        :param price: The price of this ProductInput.
        :type price: float
        """
        if price is None:
            raise ValueError("Invalid value for `price`, must not be `None`")  # noqa: E501

        self._price = price

    @property
    def description(self) -> str:
        """Gets the description of this ProductInput.

        Mô tả ngắn về sản phẩm  # noqa: E501

        :return: The description of this ProductInput.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this ProductInput.

        Mô tả ngắn về sản phẩm  # noqa: E501

        :param description: The description of this ProductInput.
        :type description: str
        """

        self._description = description